Move link to entries in entries tab
