+++ frontend/src/views/EditAry/RightPanel/styles.scss
@@ -7,11 +7,6 @@ $height-tabs: 32px;
     height: $height-tabs;
 }
 
-.entries-link {
-    padding: $spacing-small $spacing-medium;
-    color: $color-text-label;
-}
-
 .save-button {
     margin: 0 $spacing-small;
 }
